Volumes and storages refactoring:   1. Moved Volume to separate file   2. Created IVolume interface and implemented current behaviour in implementation of new interface — VolumeJBOD   3. Replaced all old volume usages with new VolumeJBOD. Where it is unnecessary to have JBOD — left just IVolume.   4. Removed old Volume completely   5. Moved StoragePolicy to separated files   6. Moved DiskSelector to separated files   7. Removed DiskSpaceMonitor file
