+++ src/Disks/IVolume.cpp
@@ -1,0 +1,46 @@
+#include "IVolume.h"
+
+#include <Common/StringUtils/StringUtils.h>
+#include <Common/escapeForFileName.h>
+#include <Common/formatReadable.h>
+#include <Common/quoteString.h>
+#include <common/logger_useful.h>
+
+#include <memory>
+
+namespace DB
+{
+namespace ErrorCodes
+{
+    extern const int EXCESSIVE_ELEMENT_IN_CONFIG;
+}
+
+IVolume::IVolume(
+    String name_, const Poco::Util::AbstractConfiguration & config, const String & config_prefix, DiskSelectorPtr disk_selector)
+    : name(std::move(name_))
+{
+    Poco::Util::AbstractConfiguration::Keys keys;
+    config.keys(config_prefix, keys);
+
+    for (const auto & disk : keys)
+    {
+        if (startsWith(disk, "disk"))
+        {
+            auto disk_name = config.getString(config_prefix + "." + disk);
+            disks.push_back(disk_selector->get(disk_name));
+        }
+    }
+
+    if (disks.empty())
+        throw Exception("Volume must contain at least one disk.", ErrorCodes::EXCESSIVE_ELEMENT_IN_CONFIG);
+}
+
+UInt64 IVolume::getMaxUnreservedFreeSpace() const
+{
+    UInt64 res = 0;
+    for (const auto & disk : disks)
+        res = std::max(res, disk->getUnreservedSpace());
+    return res;
+}
+
+}
